rename .def to .h
