+++ gd_spritestudio/shader_alpha.h
