+++ gd_spritestudio/shader_alpha_add.h
