+++ gd_spritestudio/shader_alpha_mul.h
