+++ gd_spritestudio/shader_alpha_sub.h
