+++ gd_spritestudio/shader_color.h
