+++ gd_spritestudio/shader_color_add.h
