+++ gd_spritestudio/shader_color_mix.h
