+++ gd_spritestudio/shader_color_mul.h
