+++ gd_spritestudio/shader_color_ss_sepia.h
