+++ gd_spritestudio/shader_color_sub.h
