+++ gd_spritestudio/shader_default.h
