+++ gd_spritestudio/ss_renderer_impl.cpp
@@ -17,21 +17,21 @@
 #include "ss_macros.h"
 #include "ss_texture_impl.h"
 
-#include "shader_color.def"
-#include "shader_color_ss_bmask.def"
-#include "shader_color_ss_circle.def"
-#include "shader_color_ss_hsb.def"
-#include "shader_color_ss_move.def"
-#include "shader_color_ss_noise.def"
-#include "shader_color_ss_outline.def"
-#include "shader_color_ss_pix.def"
-#include "shader_color_ss_scatter.def"
-#include "shader_color_ss_sepia.def"
-#include "shader_color_ss_spot.def"
-#include "shader_color_ss_step.def"
-#include "shader_color_ss_wave.def"
+#include "shader_color.h"
+#include "shader_color_ss_bmask.h"
+#include "shader_color_ss_circle.h"
+#include "shader_color_ss_hsb.h"
+#include "shader_color_ss_move.h"
+#include "shader_color_ss_noise.h"
+#include "shader_color_ss_outline.h"
+#include "shader_color_ss_pix.h"
+#include "shader_color_ss_scatter.h"
+#include "shader_color_ss_sepia.h"
+#include "shader_color_ss_spot.h"
+#include "shader_color_ss_step.h"
+#include "shader_color_ss_wave.h"
 
-#include "shader_alpha.def"
+#include "shader_alpha.h"
 
 SsSdkUsing
